Adds better types
