+++ src/app/proposals/sponsor/components/SponsorActions.tsx
@@ -8,6 +8,7 @@ const SponsorActions = ({
   draftProposal: DraftProposal;
 }) => {
   const renderAction = (proposal: DraftProposal) => {
+    console.log(proposal.proposal_type);
     switch (proposal.proposal_type) {
       case ProposalType.BASIC:
         return <BasicProposalAction draftProposal={proposal} />;
